// src/app/movies/interfaces/movie.interface.ts
// Generated by https://quicktype.io

export interface Result {
  dates?:         DateObj;
  page:          number;
  results:       Movie[];
  total_pages:   number;
  total_results: number;
}

export interface Movie {
  adult:             boolean;
  backdrop_path:     string;
  genres:            Genre[];
  genre_ids:        number[];
  id:                number;
  original_language: OriginalLanguage;
  original_title:    string;
  overview: string;
  tagline: string;
  popularity:        number;
  poster_path:       string;
  release_date:      string;
  title:             string;
  video:             boolean;
  vote_average:      number;
  vote_count: number;
  runtime: number;
  favorite: boolean;
}

export interface DateObj {
  minimum: string;
  maximum: string;
}

export interface Genre {
  id: number;
  name: string;
}

export enum OriginalLanguage {
  En = "en",
  Hi = "hi",
}

export interface Genres {
  genres: Genre[];
}